tests/test_p0547_number_of_provinces.py
# flake8: noqa: F403, F405
import pytest
from leetcode.p0547_number_of_provinces import *

solutions = [
    findCircaNum_bfs,
    findCircleNum_uf,
]

test_cases = [
    ([[1, 1, 0], [1, 1, 0], [0, 0, 1]], 2),
    (
        [
            [1, 1, 1, 0, 1, 1, 1, 0, 0, 0],
            [1, 1, 0, 0, 0, 0, 0, 1, 0, 0],
            [1, 0, 1, 0, 0, 0, 0, 0, 0, 0],
            [0, 0, 0, 1, 1, 0, 0, 0, 1, 0],
            [1, 0, 0, 1, 1, 0, 0, 0, 0, 0],
            [1, 0, 0, 0, 0, 1, 0, 0, 0, 0],
            [1, 0, 0, 0, 0, 0, 1, 0, 1, 0],
            [0, 1, 0, 0, 0, 0, 0, 1, 0, 1],
            [0, 0, 0, 1, 0, 0, 1, 0, 1, 1],
            [0, 0, 0, 0, 0, 0, 0, 1, 1, 1],
        ],
        1,
    ),
    ([[1, 0, 0], [0, 1, 0], [0, 0, 1]], 3),
]


@pytest.mark.timeout(1)
@pytest.mark.parametrize(("args", "expectation"), test_cases)
@pytest.mark.parametrize("solution", solutions)
def test_solution(args, expectation, solution):
    assert solution(args) == expectation
